+++ example-apps/dynamic-dropdown/triggers/species.js
@@ -13,7 +13,7 @@ const fetchList = (z, bundle) => {
   }
 
   return z.request(request).then(response => {
-    const speciesArray = JSON.parse(response.content).results;
+    const speciesArray = response.json.results;
     speciesArray.forEach(species => {
       // copy the "url" field into an "id" field
       species.id = generateID(species.url);