+++ example-apps/session-auth/authentication.js
@@ -1,11 +1,10 @@
-const testAuth = (z /*, bundle */) => {
+const testAuth = (z /*, bundle */) =>
   // Normally you want to make a request to an endpoint that is either specifically designed to test auth, or one that
   // every user will have access to, such as an account or profile endpoint like /me.
   // In this example, we'll hit httpbin, which validates the Authorization Header against the arguments passed in the URL path
-  return z.request({
+  z.request({
     url: 'https://auth-json-server.zapier-staging.com/me'
   });
-};
 
 const getSessionKey = (z, bundle) => {
   const promise = z.request({
@@ -17,12 +16,9 @@ const getSessionKey = (z, bundle) => {
     }
   });
 
-  return promise.then(response => {
-    const json = JSON.parse(response.content);
-    return {
-      sessionKey: json.sessionKey || 'secret'
-    };
-  });
+  return promise.then(response => ({
+    sessionKey: response.json.sessionKey || 'secret'
+  }));
 };
 
 module.exports = {