+++ example-apps/trigger/triggers/recipe.js
@@ -15,9 +15,7 @@ const listRecipes = (z, bundle) => {
   };
 
   // You may return a promise or a normal data structure from any perform method.
-  return z
-    .request(requestOptions)
-    .then(response => z.JSON.parse(response.content));
+  return z.request(requestOptions).then(response => response.json);
 };
 
 // We recommend writing your triggers separate like this and rolling them