+++ example-apps/typescript/src/authentication.ts
@@ -11,7 +11,7 @@ const test = async (z: ZObject /*, bundle*/) => {
   // This method can return any truthy value to indicate the credentials are valid.
   // Raise an error to show
   if (response.status === 401) {
-    throw new Error("The username and/or password you supplied is incorrect");
+    throw new z.errors.Error("The username and/or password you supplied is incorrect", "AuthenticationError", response.status);
   }
   return response;
 };