+++ packages/legacy-scripting-runner/test/integration-test.js
@@ -423,12 +423,34 @@ describe('Integration Test', () => {
     const compiledApp = schemaTools.prepareApp(appDefWithAuth);
     const app = createApp(appDefWithAuth);
 
-    it('KEY_poll', () => {
+    it('scriptingless, curlies in URL', () => {
       const appDef = _.cloneDeep(appDefinition);
       appDef.legacy.scriptingSource = appDef.legacy.scriptingSource.replace(
-        'movie_post_poll_no_id',
-        'movie_post_poll'
+        'movie_post_poll_make_array',
+        'recipe_post_poll'
       );
+      appDef.legacy.triggers.recipe.operation.url =
+        'https://httpbin.zapier-tooling.com/get?name={{name}}&active={{active}}';
+      const _compiledApp = schemaTools.prepareApp(appDef);
+      const _app = createApp(appDef);
+
+      const input = createTestInput(
+        _compiledApp,
+        'triggers.recipe.operation.perform'
+      );
+      input.bundle.authData = { name: 'john' };
+      input.bundle.inputData = { name: 'johnny', active: false };
+      return _app(input).then(output => {
+        const echoed = output.results[0];
+        should.deepEqual(echoed.args, { name: ['john'], active: ['False'] });
+        should.equal(
+          echoed.url,
+          'https://httpbin.zapier-tooling.com/get?name=john&active=False'
+        );
+      });
+    });
+
+    it('KEY_poll', () => {
       const input = createTestInput(
         compiledApp,
         'triggers.contact_full.operation.perform'
@@ -758,14 +780,22 @@ describe('Integration Test', () => {
 
     it('KEY_pre_poll, array curlies', () => {
       const appDef = _.cloneDeep(appDefinition);
-      appDef.legacy.scriptingSource = appDef.legacy.scriptingSource.replace(
-        'movie_pre_poll_array_curlies',
-        'movie_pre_poll'
-      );
+      appDef.legacy.triggers.movie.operation.url =
+        'https://httpbin.zapier-tooling.com/get?things={{things}}';
       appDef.legacy.scriptingSource = appDef.legacy.scriptingSource.replace(
         'movie_post_poll_make_array',
         'movie_post_poll'
       );
+
+      // appDef will be injected to `input` as `input._zapier.app` by
+      // createInput(). core has an injectInput beforeRequest middleware that
+      // inject `input` to the request object. When core does prepareRequest,
+      // it's going to replace every curlies in req.input._zapier.app. So apart
+      // from the request options passed into z.request, we also need to make
+      // sure array curlies from another trigger/action don't break either.
+      appDef.legacy.triggers.recipe.operation.url =
+        'https://example.com?things={{bundle.inputData.things}}';
+
       const compiledApp = schemaTools.prepareApp(appDef);
       const app = createApp(appDefWithAuth);
 
@@ -784,6 +814,30 @@ describe('Integration Test', () => {
         req.url.should.equal(
           'https://httpbin.zapier-tooling.com/get?things=eyedrops%2Ccyclops%2Cipod'
         );
+      });
+    });
+
+    it('KEY_pre_poll, GET with body', () => {
+      const appDef = _.cloneDeep(appDefinition);
+      appDef.legacy.scriptingSource = appDef.legacy.scriptingSource.replace(
+        'movie_pre_poll_GET_with_body',
+        'movie_pre_poll'
+      );
+      appDef.legacy.scriptingSource = appDef.legacy.scriptingSource.replace(
+        'movie_post_poll_make_array',
+        'movie_post_poll'
+      );
+      const _appDefWithAuth = withAuth(appDef, apiKeyAuth);
+      const _compiledApp = schemaTools.prepareApp(_appDefWithAuth);
+      const _app = createApp(_appDefWithAuth);
+
+      const input = createTestInput(
+        _compiledApp,
+        'triggers.movie.operation.perform'
+      );
+      return _app(input).then(output => {
+        const echoed = output.results[0];
+        should.equal(echoed.args.name, 'Luke Skywalker');
       });
     });
 
@@ -1490,19 +1544,24 @@ describe('Integration Test', () => {
         'triggers.contact_hook_scripting.operation.performUnsubscribe'
       );
       input.bundle.authData = { api_key: 'yo yo' };
-      input.bundle.inputData = { foo: 'bar' };
+      input.bundle.inputData = { foo: 'bar', subscription_id: 8866 };
       input.bundle.targetUrl = 'https://foo.bar';
+      input.bundle.subscribeData = { subscription_id: 7744 };
       input.bundle.meta = { zap: { id: 9512 } };
       return app(input).then(output => {
         should.equal(output.results.request.method, 'DELETE');
 
         const echoed = output.results.json;
+        should.equal(echoed.args.sub_id, '7744');
         should.equal(echoed.json.event, 'contact.created');
         should.equal(echoed.json.hidden_message, 'pre_unsubscribe was here!');
         should.equal(echoed.headers['X-Api-Key'], 'yo yo');
 
         should.deepEqual(echoed.json.bundleAuthFields, { api_key: 'yo yo' });
-        should.deepEqual(echoed.json.bundleTriggerFields, { foo: 'bar' });
+        should.deepEqual(echoed.json.bundleTriggerFields, {
+          foo: 'bar',
+          subscription_id: 8866
+        });
         should.equal(echoed.json.bundleTargetUrl, 'https://foo.bar');
         should.equal(echoed.json.bundleEvent, 'contact.created');
         should.deepEqual(echoed.json.bundleZap, { id: 9512 });