+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 9.4.0
+
+### cli
+
+* :nail_care: `build` and `push` command now produces smaller zips (≈30% of the original size!) ([#202](https://github.com/zapier/zapier-platform/pull/202))
+
+### core
+
+* :nail_care: `z.request` now has an `allowGetBody` option that allows you to send a GET request with a body ([#195](https://github.com/zapier/zapier-platform/pull/195))
+* :gear: Include `User-Agent` header for internal calls ([#204](https://github.com/zapier/zapier-platform/pull/204))
+* :scroll: Update examples to demonstrate `z.errors.Error` ([#198](https://github.com/zapier/zapier-platform/pull/198))
+* :scroll: Encourage use of `response.json` rather than `z.JSON.parse(response.content)` ([#200](https://github.com/zapier/zapier-platform/pull/200))
+
+### schema
+
+* No changes
+
 ## 9.3.0
 
 ### cli